Refactor namespace moves
